+++ include/snd/audio/filter/1-pole_mono.h
@@ -8,7 +8,7 @@ namespace filter {
 
 class Filter_1Pole_Mono
 {
-	float sr_;
+	int sr_;
 	float freq_;
 
 	Filter_1Pole filter_;
@@ -21,7 +21,7 @@ public:
 	float hp() const { return filter_.hp(); }
 	void process_frame(float in);
 	void set_freq(float freq, bool recalculate = true);
-	void set_sr(float sr, bool recalculate = true);
+	void set_sr(int sr, bool recalculate = true);
 	void recalculate();
 };
