+++ include/snd/audio/filter/2-pole_stereo.h
@@ -8,7 +8,7 @@ namespace filter {
 
 class Filter_2Pole_Stereo
 {
-	float sr_;
+	int sr_;
 	float freq_;
 	float res_;
 
@@ -31,7 +31,7 @@ public:
 	void process_right(float in);
 	void set_freq(float freq, bool recalculate = true);
 	void set_res(float res, bool recalculate = true);
-	void set_sr(float sr, bool recalculate = true);
+	void set_sr(int sr, bool recalculate = true);
 	void recalculate();
 };
