+++ include/snd/audio/ringmod/ring_modulator_mono.h
@@ -8,7 +8,7 @@ namespace ringmod {
 
 class RingModulator_Mono
 {
-	float sr_ = 44100.0f;
+	int sr_ = 44100;
 	float freq_ = 100.0f;
 	float amount_ = 0.0f;
 
@@ -18,7 +18,7 @@ public:
 
 	float operator()(float in);
 	void reset(float phase);
-	void set_sr(float sr, bool recalculate = true);
+	void set_sr(int sr, bool recalculate = true);
 	void set_freq(float freq, bool recalculate = true);
 	void set_amount(float amount);
 	void recalculate();
