+++ src/audio/filter/2-pole_stereo.cpp
@@ -6,7 +6,7 @@ namespace audio {
 namespace filter {
 
 Filter_2Pole_Stereo::Filter_2Pole_Stereo()
-	: sr_(44100.f)
+	: sr_(44100)
 	, freq_(1.f)
 	, res_(0.f)
 {
@@ -36,7 +36,7 @@ void Filter_2Pole_Stereo::set_res(float res, bool recalc)
 	if (recalc)	recalculate();
 }
 
-void Filter_2Pole_Stereo::set_sr(float sr, bool recalc)
+void Filter_2Pole_Stereo::set_sr(int sr, bool recalc)
 {
 	sr_ = sr;
